enviando informacoes para o firebase

+++ app/src/main/java/br/com/truckhospital/modules/core/model/Budget.kt
@@ -3,7 +3,7 @@ package br.com.truckhospital.modules.core.model
 import java.io.Serializable
 
 data class Budget(
-    val laborCost: Float,
-    val partsCost: Float,
-    val totalCost: Float
+    val laborCost: Double,
+    val partsCost: Double,
+    val totalCost: Double
 ) : Serializable
